Change type to float so we can support NaN nid values

+++ hatchet/tests/caliper.py
@@ -195,7 +195,7 @@ def test_graphframe_native_lulesh_from_file(lulesh_caliper_cali):
         if col in ("time (inc)", "time"):
             assert gf.dataframe[col].dtype == np.float64
         elif col in ("nid", "rank"):
-            assert gf.dataframe[col].dtype == np.int64
+            assert gf.dataframe[col].dtype == np.float64
         elif col in ("name", "node"):
             assert gf.dataframe[col].dtype == object
 
@@ -221,7 +221,7 @@ def test_graphframe_native_lulesh_from_caliperreader(lulesh_caliper_cali):
         if col in ("time (inc)", "time"):
             assert gf.dataframe[col].dtype == np.float64
         elif col in ("nid", "rank"):
-            assert gf.dataframe[col].dtype == np.int64
+            assert gf.dataframe[col].dtype == np.float64
         elif col in ("name", "node"):
             assert gf.dataframe[col].dtype == object
 
@@ -881,7 +881,7 @@ def test_graphframe_timeseries_lulesh_from_file(caliper_timeseries_cali):
         if col in ("time (inc)", "time"):
             assert gf.dataframe[col].dtype == np.float64
         elif col in ("nid", "rank"):
-            assert gf.dataframe[col].dtype == np.int64
+            assert gf.dataframe[col].dtype == np.float64
         elif col in ("name", "node"):
             assert gf.dataframe[col].dtype == object
